requirements.txt
Pillow==7.2.0
torch==1.6.0
torchvision==0.7.0